# docs/swagger.yaml
basePath: /
definitions:
...
  api.CreateTextRequest:
    properties:
      metadata:
        type: string
      name:
        type: string
      text:
        type: string
    type: object
...
  api.LoginRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  api.LoginResponse:
    properties:
      access_token:
        type: string
      user_id:
        type: integer
    type: object
  api.RegisterRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
...
  api.TextItemResponse:
    properties:
      checksum:
        type: string
      created_at:
        type: string
      metadata:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  api.TextResponse:
    properties:
      checksum:
        type: string
      created_at:
        type: string
      metadata:
        type: string
      name:
        type: string
      text:
        type: string
      updated_at:
        type: string
    type: object
...
  api.UpdateTextRequest:
    properties:
      metadata:
        type: string
      text:
        type: string
    type: object
info:
  contact: {}
  description: Менеджер паролей GophKeeper
  title: GophKeeper API
  version: "1.0"
paths:
...
  /api/text:
    get:
      description: Возвращает имена и метаданные текстов текущего пользователя без
        их содержимого
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.TextItemResponse'
            type: array
        "401":
          description: Пользователь не аутентифицирован
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Список текстовых данных пользователя
      tags:
      - text
    post:
      consumes:
      - application/json
      description: Сохраняет произвольные текстовые данные текущего пользователя.
        Имя уникально в рамках пользователя
      parameters:
      - description: Текстовые данные
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.CreateTextRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TextResponse'
        "400":
          description: Неверный формат запроса
          schema:
            type: string
        "401":
          description: Пользователь не аутентифицирован
          schema:
            type: string
        "409":
          description: Текст с таким именем уже существует
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Создание текстовых данных
      tags:
      - text
  /api/text/{text_name}:
    delete:
      description: Удаляет текстовые данные текущего пользователя по имени
      parameters:
      - description: Имя текста
        in: path
        name: text_name
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "204":
          description: No Content
        "400":
          description: Неверный формат запроса
          schema:
            type: string
        "401":
          description: Пользователь не аутентифицирован
          schema:
            type: string
        "404":
          description: Текст не найден
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удаление текстовых данных
      tags:
      - text
    get:
      description: Возвращает текстовые данные текущего пользователя по имени
      parameters:
      - description: Имя текста
        in: path
        name: text_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TextResponse'
        "400":
          description: Неверный формат запроса
          schema:
            type: string
        "401":
          description: Пользователь не аутентифицирован
          schema:
            type: string
        "404":
          description: Текст не найден
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получение текстовых данных
      tags:
      - text
    put:
      consumes:
      - application/json
      description: Полностью перезаписывает текстовые данные текущего пользователя,
        найденные по имени
      parameters:
      - description: Имя текста
        in: path
        name: text_name
        required: true
        type: string
      - description: Новые текстовые данные
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateTextRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TextResponse'
        "400":
          description: Неверный формат запроса
          schema:
            type: string
        "401":
          description: Пользователь не аутентифицирован
          schema:
            type: string
        "404":
          description: Текст не найден
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Обновление текстовых данных
      tags:
      - text
  /api/user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные пользователя
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LoginResponse'
        "400":
          description: Неверный формат запроса
          schema:
            type: string
        "401":
          description: Неверный логин или пароль
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      summary: Аутентификация пользователя
      tags:
      - auth
  /api/user/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные пользователя
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.RegisterRequest'
      produces:
      - text/plain
      responses:
        "200":
          description: OK
        "400":
          description: Неверный формат запроса или недопустимая длина логина либо
            пароля
          schema:
            type: string
        "409":
          description: Логин уже занят
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      summary: Регистрация пользователя
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"